Moved js code for login into it's own file for cleaner coding practices
